First commit. Basic dice classes created.
Dice: Trait which defines basic dice behaviour.
SuccessfulDice: Trait which defines some extra behaviours to check the
result of a dice rolling.
AbstractDice: Abstract class which implements common functionality.
SimpleDice: A simple numerical dice.
SimpleSuccessfulDice: A numerical dice which can check if it's a
successful rolling or not, depending on a numerical value considered as
difficult.
ReverseSuccessfulDice: A SimpleSuccessfulDice which has a successful
result if it's <= difficult.
SimpleCommandDice: A SimpleSuccessfulDice which can execute some action
on success or fail.
GenericDice: A dice which can have anything as sides.
GenericSuccessfulDice: A GenericDice which can mark some sides as
successful results.
GenericCommandDice: A GenericSuccessfulDice which can execute some
actions on success or error.